+++ src/Data/Managers/V2/DesiredState/GetDesiredStates/getUrl.ts
@@ -1,7 +1,7 @@
 import moment from "moment-timezone";
 import qs from "qs";
 import { RangeOperator } from "@/Core";
-import { urlEncodeParams } from "../../helpers/utils";
+import { urlEncodeParams } from "../../helpers";
 import { GetDesiredStatesParams } from "./useGetDesiredStates";
 
 /**
