+++ src/Data/Managers/V2/Notification/GetNotifications/getUrl.ts
@@ -1,5 +1,5 @@
 import qs from "qs";
-import { urlEncodeParams } from "../../helpers/utils";
+import { urlEncodeParams } from "../../helpers";
 import { GetNotificationsParams } from "./useGetNotifications";
 
 export function getUrl(params: GetNotificationsParams): string {
