+++ src/Data/Managers/V2/Resource/GetResourceLogs/getUrl.ts
@@ -1,7 +1,7 @@
 import moment from "moment-timezone";
 import qs from "qs";
 import { DateRange, RangeOperator } from "@/Core";
-import { urlEncodeParams } from "../../helpers/utils";
+import { urlEncodeParams } from "../../helpers";
 import { GetResourceLogsParams } from "./useGetResourceLogs";
 
 /**
