+++ src/Data/Managers/V2/Resource/GetResources/getUrl.ts
@@ -1,5 +1,5 @@
 import qs from "qs";
-import { urlEncodeParams } from "../../helpers/utils";
+import { urlEncodeParams } from "../../helpers";
 import { GetResourcesParams } from "./useGetResources";
 
 /**
