--- src/Data/Managers/V2/helpers/utils.ts
@@ -1,11 +1,0 @@
-export function urlEncodeParams<Type extends object>(params: Type) {
-  const encodedQuery = Object.fromEntries(
-    Object.entries(params).map(([key, value]) => {
-      const encodedValue = typeof value === "string" ? encodeURIComponent(value) : value;
-
-      return [key, encodedValue];
-    })
-  );
-
-  return encodedQuery as Type;
-}
